+++ awslimitchecker/tests/services/result_fixtures.py
@@ -2450,7 +2450,10 @@ class ApiGateway(object):
                     ],
                     'binaryMediaTypes': [
                         'string',
-                    ]
+                    ],
+                    'endpointConfiguration': {
+                        'types': ['PRIVATE']
+                    }
                 },
                 {
                     'id': 'api2',
@@ -2459,7 +2462,10 @@ class ApiGateway(object):
                     'createdDate': datetime(2016, 1, 1),
                     'version': 'api2ver',
                     'warnings': [],
-                    'binaryMediaTypes': []
+                    'binaryMediaTypes': [],
+                    'endpointConfiguration': {
+                        'types': ['REGIONAL']
+                    }
                 }
             ],
             'NextToken': 'string'
@@ -2477,8 +2483,43 @@ class ApiGateway(object):
                     ],
                     'binaryMediaTypes': [
                         'string',
-                    ]
+                    ],
+                    'endpointConfiguration': {
+                        'types': ['REGIONAL']
+                    }
                 },
+                {
+                    'id': 'api4',
+                    'name': 'api4name',
+                    'description': 'api4desc',
+                    'createdDate': datetime(2017, 1, 2),
+                    'version': 'api4ver',
+                    'warnings': [
+                        'string',
+                    ],
+                    'binaryMediaTypes': [
+                        'string',
+                    ],
+                    'endpointConfiguration': {
+                        'types': ['EDGE']
+                    }
+                },
+                {
+                    'id': 'api5',
+                    'name': 'api5name',
+                    'description': 'api5desc',
+                    'createdDate': datetime(2017, 1, 2),
+                    'version': 'api5ver',
+                    'warnings': [
+                        'string',
+                    ],
+                    'binaryMediaTypes': [
+                        'string',
+                    ],
+                    'endpointConfiguration': {
+                        'types': ['EDGE']
+                    }
+                }
             ]
         }
     ]
@@ -2829,10 +2870,16 @@ class ApiGateway(object):
 
     resources_api3 = [{'items': []}]
 
+    resources_api4 = [{'items': []}]
+
+    resources_api5 = [{'items': []}]
+
     get_resources = {
         'api1': resources_api1,
         'api2': resources_api2,
-        'api3': resources_api3
+        'api3': resources_api3,
+        'api4': resources_api4,
+        'api5': resources_api5
     }
 
     doc_parts = {
@@ -2918,7 +2965,33 @@ class ApiGateway(object):
                 },
                 'properties': 'string'
             }
-        ]
+        ],
+        'api4': [
+            {
+                'id': 'string',
+                'location': {
+                    'type': 'API',
+                    'path': 'string',
+                    'method': 'string',
+                    'statusCode': 'string',
+                    'name': 'string'
+                },
+                'properties': 'string'
+            }
+        ],
+        'api5': [
+            {
+                'id': 'string',
+                'location': {
+                    'type': 'API',
+                    'path': 'string',
+                    'method': 'string',
+                    'statusCode': 'string',
+                    'name': 'string'
+                },
+                'properties': 'string'
+            }
+        ],
     }
 
     stages = {
@@ -2968,6 +3041,16 @@ class ApiGateway(object):
             'item': [
                 {'deploymentId': 'blam'},
                 {'deploymentId': 'blarg'}
+            ]
+        },
+        'api4': {
+            'item': [
+                {'deploymentId': 'baz'}
+            ]
+        },
+        'api5': {
+            'item': [
+                {'deploymentId': 'baz'}
             ]
         }
     }
@@ -3019,7 +3102,9 @@ class ApiGateway(object):
                 'authorizerResultTtlInSeconds': 123
             }
         ],
-        'api3': []
+        'api3': [],
+        'api4': [],
+        'api5': []
     }
 
     plans = [
@@ -3140,6 +3225,35 @@ class ApiGateway(object):
                     'createdDate': datetime(2015, 1, 1),
                     'expirationDate': datetime(2015, 1, 1)
                 },
+            ]
+        }
+    ]
+
+    vpc_links = [
+        {
+            'items': [
+                {
+                    'id': 'vpcl-1',
+                    'name': 'link1',
+                    'description': 'desc1',
+                    'status': 'AVAILABLE'
+                }
+            ]
+        },
+        {
+            'items': [
+                {
+                    'id': 'vpcl-2',
+                    'name': 'link2',
+                    'description': 'desc2',
+                    'status': 'AVAILABLE'
+                },
+                {
+                    'id': 'vpcl-3',
+                    'name': 'link3',
+                    'description': 'desc3',
+                    'status': 'PENDING'
+                }
             ]
         }
     ]